requirements.txt
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backoff==2.2.1
certifi==2022.12.7
cffi==1.15.1
# channels-redis==3.3.1
charset-normalizer==3.1.0
cli-exit-tools==1.2.3.2
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==1.3.0
ecdsa==0.18.0
ffmpeg==1.4
frozenlist==1.3.3
gunicorn==20.1.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jwcrypto==1.4.2
lib-detect-testenv==2.0.3
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.12.2
oauthlib==3.2.2
openai==0.27.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
# redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
tqdm==4.65.0
# Twisted==22.10.0
# twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wrapt==1.15.0
wrapt-timeout-decorator==1.3.12.2
yarl==1.8.2
zope.interface==5.5.2
